Public API

IsDef.:∨

IsDef.Out

IsDef.apply

IsDef.isdef

IsDef.return_type

Core

IsDef.isdef — Function

checks whether the function is defined for the actual types or not

This works in compile time and hence can be used to optimize code.

IMPORTANT: Overload $IsDef.return_type$ if you experience unexpected behaviour for your types For instance to say that some call like $myfunc(::Int, ::String)$ is not defined define the following

function IsDef.return_type(::Type{Tuple{typeof(myfunc), TypeArg1, TypeArg2}})
  Union{}  # return empty Union to indicate something is not defined
end

IsDef.Out — Function

returns outputtype of function application

Returns $Traits.NotApplicable$ if compiler notices that no Method can be found

CAUTION: If $Out(...) == Any$, still a MethodError might happen at runtime. This is due to incomplete type inference.

SOLUTION: Overload $IsDef.return_type$ if you experience unexpected behaviour for your types For instance to say that some call like $myfunc(::Int, ::String)$ is not defined, then define the following

function IsDef.return_type(::Type{Tuple{typeof(myfunc), TypeArg1, TypeArg2}})
  Union{}  # return empty Union to indicate something is not defined
end

IsDef.return_type — Function

wrapper arround Julia's type inference

This should be overloaded if you want to fix certain wrong typeinferences for your custom types. Returning $Union{}$ is interpreted as $MethodError$.

It is used internally by both $isdef$ and $Out$.

Other helpers

IsDef.apply — Function

just applies a given function to arguments and keyword arguments

This little helper is crucial if you want to typeinfer when only knowing the function type instead of the function instance.

IsDef.:∨ — Function

∨ (latex ee) is alias for $promote_type$

when called on values, the values will be cast to types via use of typeof for convenience