Public API

IsDef.:∨

IsDef.Out

IsDef.apply

IsDef.isdef

IsDef.return_type

Core

IsDef.isdef — Function

isdef(func, ArgType1, ArgType2, ...)::Bool
isdef(f, args...) = isdef(f, typeof.(args)...)

Checks whether the function is defined for the actual types or not. This works in compile time and hence can be used to optimize code.

When called on values, the values will be cast to types via use of typeof for convenience.

CAUTION

If isdef(...) == true, still a MethodError might happen at runtime. This is due to incomplete type inference.

SOLUTION: Overload IsDef.return_type if you experience unexpected behaviour for your types

For instance to say that some call like myfunc(::Int, ::String) is not defined, then define the following

function IsDef.return_type(::Type{Tuple{typeof(myfunc), TypeArg1, TypeArg2}})
  Union{}  # return empty Union to indicate something is not defined
end

Semantics

The implementation follows a closed-world semantics in that isdef will return true for a given abstract type, if the function is defined for all currently defined leaf types of that abstract type.

The only exceptions currently are Any, Function, and Exception, for which isdef returns true.

IsDef.Out — Function

Out(func, ArgType1, ArgType2, ...)::ReturnType
Out(f, args...) = isdef(f, typeof.(args)...)

Returns outputtype of function application. Returns Traits.NotApplicable if compiler notices that no Method can be found.

When called on values, the values will be cast to types via use of typeof for convenience.

CAUTION

If Out(...) == Any, still a MethodError might happen at runtime. This is due to incomplete type inference.

SOLUTION: Overload IsDef.return_type if you experience unexpected behaviour for your types

For instance to say that some call like myfunc(::Int, ::String) is not defined, then define the following

function IsDef.return_type(::Type{Tuple{typeof(myfunc), TypeArg1, TypeArg2}})
  Union{}  # return empty Union to indicate something is not defined
end

UnionAll types and abstract types are concretified to the Union of their existing subtypes, in order to improve type-inference. The only exceptions so far are Any, Function and Exception, as they have way too many subtypes to be of practical use.

IsDef.return_type — Function

wrapper arround Julia's type inference

This should be overloaded if you want to fix certain wrong typeinferences for your custom types. It is used internally by both isdef and Out.

Returning Union{} is interpreted as MethodError.

Other helpers

IsDef.apply — Function

just applies a given function to arguments and keyword arguments

This little helper is crucial if you want to typeinfer when only knowing the function type instead of the function instance.

IsDef.:∨ — Function

TypeA ∨ TypeB = promote_type(TypeA, TypeB)
∨(values...) = ∨(typeof.(values)...)

∨ (latex ee) is alias for promote_type.

When called on values, the values will be cast to types via use of typeof for convenience.