IsDef

IsDef.IsDef

IsDef.:∨

IsDef.Out

IsDef._return_type

IsDef.apply

IsDef.isdef

IsDef.@redefine_generated

IsDef.IsDef — Module

To more easily dispatch on the availability of functions, this package offers two essential helpers:

`isdef(f, TypeArg1, TypeArg2, ...)` and `Out(f, TypeArg1, TypeArg2, ...)`

Currently the implementation follows a closed-world semantics in that $isdef$ will return true for a given abstract type, if the function is defined for all currently defined leaf types of that abstract type.

The only exception is Any, for which isdef returns true only if given an abstract type, the function is defined for a newtype of the abstract type. (E.g. $f(a) = 1$ )

IsDef.:∨ — Function

∨ (latex ee) is alias for $promote_type$

when called on values, the values will be cast to types via use of typeof for convenience

IsDef.Out — Method

returns outputtype of function application

Returns $Traits.NotApplicable$ if compiler notices that no Method can be found

CAUTION: If $Out(...) == Any$, still a MethodError might happen at runtime. This is due to incomplete type inference.

SOLUTION: Overload $IsDef.return_type$ if you experience unexpected behaviour for your types For instance to say that some call like $myfunc(::Int, ::String)$ is not defined, then define the following

function IsDef.return_type(::Type{Tuple{typeof(myfunc), TypeArg1, TypeArg2}})
  Union{}  # return empty Union to indicate something is not defined
end

IsDef._return_type — Method

_return_type can either except

a Union (which may be created by Type2Union)

_ or a Tuple (which will be the final call signature)

Importantly, the semantics is that if at least one of the Union types infers Union{} then Union{} is inferred in total.

IsDef.apply — Method

just applies a given function to arguments and keyword arguments

This little helper is crucial if you want to typeinfer when only knowing the function type instead of the function instance.

IsDef.isdef — Method

checks whether the function is defined for the actual types or not

This works in compile time and hence can be used to optimize code.

IMPORTANT: Overload $IsDef.return_type$ if you experience unexpected behaviour for your types For instance to say that some call like $myfunc(::Int, ::String)$ is not defined define the following

function IsDef.return_type(::Type{Tuple{typeof(myfunc), TypeArg1, TypeArg2}})
  Union{}  # return empty Union to indicate something is not defined
end

IsDef.@redefine_generated — Macro

for some reasons the @generated macro triggers too early and leads Union{} for some types

calling this macro will redefine the @generated functions accordingly so that new TypeDefinitions are properly included